AMDGPU: return uint64_t from longfield
longfield computes its mask at 64 bits but returned int, silently
truncating any extraction wider than 32 bits. No current field is
wider than 32 (LITERAL, whose bit pattern survived the round-trip),
so this is behavior-neutral; it removes the trap for a future drop
that defines a wider field.

+++ instructionAPI/src/AMDGPU/gfx940/InstructionDecoder-amdgpu-gfx940.h
@@ -107,7 +107,7 @@ namespace Dyninst {
                 }
 
             template<int start, int end>
-                int longfield(unsigned long long int raw)  {
+                uint64_t longfield(unsigned long long int raw)  {
 #if defined DEBUG_FIELD
                     std::cerr << start << "-" << end << ":" << std::dec << (raw >> (start) &
                             (0xFFFFFFFFFFFFFFFF >> (63 - (end - start)))) << " ";
